put all sub-commands to one class
