--- src/java/org/apache/cassandra/tools/nodetool/clustermetadata/InitializeCMS.java
@@ -1,40 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.cassandra.tools.nodetool.clustermetadata;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import io.airlift.airline.Command;
-import io.airlift.airline.Option;
-import org.apache.cassandra.tools.NodeProbe;
-import org.apache.cassandra.tools.NodeTool;
-
-@Command(name = "initialize", description = "Upgrade from gossip and initialize CMS")
-public class InitializeCMS extends NodeTool.NodeToolCmd
-{
-    @Option(title = "ignored endpoints", name = {"-i", "--ignore"}, description = "Hosts to ignore due to them being down")
-    private List<String> endpoint = new ArrayList<>();
-
-    @Override
-    protected void execute(NodeProbe probe)
-    {
-        probe.getCMSOperationsProxy().initializeCMS(endpoint);
-    }
-}
